board prints, can select coord and replace with symbol, can alternate symbols
